# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog",
        "description": "Blog API Document",
        "contact": {
            "email": "tech@area02.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "抓取所有部落格文章",
                "description": "抓取所有部落格文章(不包含私人)",
                "operationId": "e93214eff756c07fefd10a2a8781abf0",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "頁碼",
                        "required": false
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "每頁幾筆",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "排序欄位(id: post.id, createdAt: 新增時間)",
                        "required": false
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "排序方式(asc: 升冪排序, desc: 降冪排序)",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts": {
                                            "description": "文章",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Posts"
                                            }
                                        },
                                        "total": {
                                            "description": "文章總數",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "新增部落格文章",
                "description": "新增部落格文章",
                "operationId": "dccb43378a995a48a4843a0b3f10f286",
                "requestBody": {
                    "description": "Posts",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "content",
                                        "type": "string"
                                    },
                                    "isAvailable": {
                                        "description": "是否公開",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Posts"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/api/posts/{id}/disable": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "作者將文章改成私人",
                "description": "作者將文章改成私人",
                "operationId": "081b0f51ef230c9d32193fc32a54013d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "文章id(post.id)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Posts"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "作者更新部落格文章",
                "description": "作者更新部落格文章",
                "operationId": "929614537d76a081a92170a41ed3bdb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "文章id(posts.id)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Posts",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "content",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Posts"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Comments": {
                "properties": {
                    "content": {
                        "description": "content",
                        "type": "string"
                    },
                    "userName": {
                        "description": "評論者名字",
                        "type": "string"
                    },
                    "isAvailable": {
                        "description": "是否公開",
                        "type": "boolean"
                    },
                    "createdAt": {
                        "description": "新增時間",
                        "type": "string",
                        "format": "timestamp",
                        "example": "1721277196"
                    },
                    "updatedAt": {
                        "description": "更新時間",
                        "type": "string",
                        "format": "timestamp",
                        "example": "1721277196"
                    }
                },
                "type": "object"
            },
            "Posts": {
                "properties": {
                    "content": {
                        "description": "content",
                        "type": "string"
                    },
                    "userName": {
                        "description": "作者名字",
                        "type": "string"
                    },
                    "isAvailable": {
                        "description": "是否公開",
                        "type": "boolean"
                    },
                    "comments": {
                        "description": "評論",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comments"
                        }
                    },
                    "createdAt": {
                        "description": "新增時間",
                        "type": "string",
                        "format": "timestamp",
                        "example": "1721277196"
                    },
                    "updatedAt": {
                        "description": "更新時間",
                        "type": "string",
                        "format": "timestamp",
                        "example": "1721277196"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "400": {
                "description": "客戶端錯誤",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "message": "客戶端錯誤",
                            "status": 400
                        }
                    }
                }
            },
            "401": {
                "description": "身份驗證失敗",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "message": "Unauthenticated",
                            "status": 401
                        }
                    }
                }
            },
...
            "500": {
                "description": "伺服器發生錯誤",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "message": "伺服器發生錯誤",
                            "status": 500
                        }
                    }
                }
            }
        }
    },
...
}